feat: 모임관련 인가로직, 모임에 가입한 회원 목록 조회 기능 구현 (#136)
* feat: 팀 조회시 joined 컬럼을 포함하도록 변경

* feat: 모임에 가입한 회원 목록 조회 기능 구현 및 테스트 작성

* feat: 팀에 가입된 사람만 팀의 정보읽기 및 수정삭제 가능 하도록 인가로직 추가

* style: 스타일 포맷터 적용

* feat: 없는 팀에 요청시 404 말고 403 뜨던거 수정

* feat: 팀 상세조회 joined 추가

* test: 변수명 변경

+++ backend/src/main/java/com/woowacourse/naepyeon/controller/TeamController.java
@@ -5,7 +5,9 @@ import com.woowacourse.naepyeon.controller.dto.CreateResponse;
 import com.woowacourse.naepyeon.controller.dto.JoinTeamMemberRequest;
 import com.woowacourse.naepyeon.controller.dto.LoginMemberRequest;
 import com.woowacourse.naepyeon.controller.dto.TeamRequest;
+import com.woowacourse.naepyeon.exception.UncertificationTeamMemberException;
 import com.woowacourse.naepyeon.service.TeamService;
+import com.woowacourse.naepyeon.service.dto.JoinedMembersResponseDto;
 import com.woowacourse.naepyeon.service.dto.TeamResponseDto;
 import com.woowacourse.naepyeon.service.dto.TeamsResponseDto;
 import java.net.URI;
@@ -32,7 +34,7 @@ public class TeamController {
     public ResponseEntity<TeamResponseDto> getTeam(
             @AuthenticationPrincipal @Valid final LoginMemberRequest loginMemberRequest,
             @PathVariable final Long teamId) {
-        return ResponseEntity.ok(teamService.findById(teamId));
+        return ResponseEntity.ok(teamService.findById(teamId, loginMemberRequest.getId()));
     }
 
     @GetMapping("/me")
@@ -42,9 +44,21 @@ public class TeamController {
     }
 
     @GetMapping
-    public ResponseEntity<TeamsResponseDto> getAllTeams() {
-        return ResponseEntity.ok(teamService.findAll());
+    public ResponseEntity<TeamsResponseDto> getAllTeams(
+            @AuthenticationPrincipal @Valid final LoginMemberRequest loginMemberRequest) {
+        return ResponseEntity.ok(teamService.findAll(loginMemberRequest.getId()));
     }
+
+    @GetMapping("/{teamId}/members")
+    public ResponseEntity<JoinedMembersResponseDto> getJoinedMembers(
+            @AuthenticationPrincipal @Valid final LoginMemberRequest loginMemberRequest,
+            @PathVariable final Long teamId) {
+        if (!teamService.isJoinedMember(loginMemberRequest.getId(), teamId)) {
+            throw new UncertificationTeamMemberException(teamId, loginMemberRequest.getId());
+        }
+        return ResponseEntity.ok(teamService.findJoinedMembers(teamId));
+    }
+
 
     @PostMapping
     public ResponseEntity<CreateResponse> createTeam(
@@ -58,6 +72,9 @@ public class TeamController {
     public ResponseEntity<Void> updateTeam(@AuthenticationPrincipal @Valid final LoginMemberRequest loginMemberRequest,
                                            @PathVariable final Long teamId,
                                            @RequestBody @Valid final TeamRequest teamRequest) {
+        if (!teamService.isJoinedMember(loginMemberRequest.getId(), teamId)) {
+            throw new UncertificationTeamMemberException(teamId, loginMemberRequest.getId());
+        }
         teamService.updateName(teamId, teamRequest.getName());
         return ResponseEntity.noContent().build();
     }
@@ -65,6 +82,9 @@ public class TeamController {
     @DeleteMapping("/{teamId}")
     public ResponseEntity<Void> deleteTeam(@AuthenticationPrincipal @Valid final LoginMemberRequest loginMemberRequest,
                                            @PathVariable final Long teamId) {
+        if (!teamService.isJoinedMember(loginMemberRequest.getId(), teamId)) {
+            throw new UncertificationTeamMemberException(teamId, loginMemberRequest.getId());
+        }
         teamService.delete(teamId);
         return ResponseEntity.noContent().build();
     }
